Implement slightly simplified case of assembly tree
- made different cases from binary 2 ops, and binary 1op + immmediate
- fixed bugs in evaluation order, added test cases
- implement left and right shifts
- do not delete files when compiling with -g option

+++ main.c
@@ -232,11 +232,13 @@ int main(int argc, char *argv[]) {
 
 cleanup_o_file:
 
-    unlink("out.o");
+    if (!debug)
+        unlink("out.o");
 
 cleanup_s_file:
 
-    unlink("out.s");
+    if (!debug)
+        unlink("out.s");
 
 cleanup_cpp_file:
